% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_by_date.R
\name{cut_by_date}
\alias{cut_by_date}
\title{Cut a dataset for analysis at a specified date}
\usage{
cut_by_date(data, cut_time)
}
\arguments{
\item{data}{A time-to-event dataset returned from \link{simu_enrich_trial}.}

\item{cut_time}{Date relative to start of randomization (\code{calendarTime} from
input dataset) at which dataset is to be cut off for analysis.}
}
\value{
A data frame ready for survival analysis including columns \code{calendarCutoff},
\code{survTimeCut}, \code{eventCut}.
}
\description{
Cut a dataset for analysis at a specified date
}
\examples{
d <- simu_enrich_trial(n = 100, prop_S = 0.5, ratio = 1)
dcut <- cut_by_date(d, 10)
}
