% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunnett.R
\name{dunnett}
\alias{dunnett}
\title{Adjusted p value for testing H_J using Dunnett method}
\usage{
dunnett(p, cr)
}
\arguments{
\item{p}{A vector of individual raw p values in J.}

\item{cr}{correlation structure.}
}
\value{
Adjusted p values using Dunnett procedure
}
\description{
This functions provides the adjusted p value for testing a family-wise hypothesis H_J based on the p values of individual raw p values in J.
}
\examples{
p = c(0.01, 0.02, 0.03, 0.013)
cr <- matrix(0.5, 4, 4); diag(cr) <- 1
dunnett(p, cr)
}
