% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDesignParameters.R
\name{getDesignParameters}
\alias{getDesignParameters}
\title{Get hazard ratio thresholds for adaptive enrichment designs}
\usage{
getDesignParameters(
  n = 800,
  prop_S = 0.5,
  duration = 20,
  targetEvents.S = c(112, 280),
  CP.Sc.threshold = 0.3,
  CP.S.threshold = 0.8,
  CP.F.threshold = 0.8,
  hazard_S = c(0.1, 0.07),
  hazard_Sc = c(0.11, 0.088),
  dropout_S = c(0, 0),
  dropout_Sc = c(0, 0),
  ratio = 1,
  alpha = 0.025
)
}
\arguments{
\item{n}{total number of subjects for full population.}

\item{prop_S}{proportion for sub-population group S.}

\item{duration}{enrollment duration in months.}

\item{targetEvents.S}{The target number of events in S; currently only support length of 2, where
futility or population selection will be given at IA without efficacy testing.}

\item{CP.Sc.threshold}{conditional power threshold for futility.}

\item{CP.S.threshold}{conditional power threshold for S.}

\item{CP.F.threshold}{conditional power threshold for F.}

\item{hazard_S}{hazard rates (h_control, h_treatment) for S}

\item{hazard_Sc}{hazard rates (h_control, h_treatment) for Sc}

\item{dropout_S}{dropout hazard rates (h_control, h_treatment) for S}

\item{dropout_Sc}{dropout hazard rates (h_control, h_treatment) for Sc}

\item{ratio}{randomization ratio r:1, where r refers to treatment arm; for equal randomization, r=1.}

\item{alpha}{Type I error, always one-sided.}
}
\value{
It returns a list.
}
\description{
Get hazard ratio thresholds for adaptive enrichment designs
}
\examples{
getDesignParameters()
}
