% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOC_par.R
\name{getOC_par}
\alias{getOC_par}
\title{Get operating characteristics via simulations for an enrichment design}
\usage{
getOC_par(
  ncore = 2,
  seed = 2024,
  nsim = 1000,
  n = 800,
  prop_S = 0.5,
  duration = 25,
  targetEvents.S = c(112, 280),
  HR.Sc.threshold = 1,
  HR.S.threshold = 0.8,
  HR.F.threshold = 0.8,
  hazard_S = c(0.1, 0.1),
  hazard_Sc = c(0.12, 0.12),
  dropout_S = c(0, 0),
  dropout_Sc = c(0, 0),
  w = 1,
  ratio = 1,
  alpha = 0.025
)
}
\arguments{
\item{ncore}{number of cores for parallel computing.}

\item{seed}{random seed for reproducibility; each core will receive a seed in \code{(1:ncore)*seed}}

\item{nsim}{number of replicates}

\item{n}{total number of subjects for full population.}

\item{prop_S}{proportion for sub-population group S.}

\item{duration}{enrollment duration in months.}

\item{targetEvents.S}{The target number of events in S; currently only support length of 2, where
futility or population selection will be given at IA without efficacy testing.}

\item{HR.Sc.threshold}{Hazard ratio threshold for futility.}

\item{HR.S.threshold}{Hazard ratio threshold for S.}

\item{HR.F.threshold}{Hazard ratio threshold for F.}

\item{hazard_S}{hazard rates (h_control, h_treatment) for S}

\item{hazard_Sc}{hazard rates (h_control, h_treatment) for Sc}

\item{dropout_S}{dropout hazard rates (h_control, h_treatment) for S}

\item{dropout_Sc}{dropout hazard rates (h_control, h_treatment) for Sc}

\item{w}{weight parameter in cumulative enrollment pattern.}

\item{ratio}{randomization ratio r:1, where r refers to treatment arm; for equal randomization, r=1.}

\item{alpha}{Type I error, always one-sided.}
}
\value{
It returns a list.
}
\description{
Get operating characteristics via simulations for an enrichment design
}
\examples{
res <- getOC_par(ncore = 2, seed = 24232, nsim=10)
lapply(res, function(x) mean(apply(x[,1:2], 1, any, na.rm=TRUE)))
}
