% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOC_ph23.R
\name{getOC_ph23}
\alias{getOC_ph23}
\title{Get operating characteristics via simulations for the phase 2/3 design}
\usage{
getOC_ph23(
  ith = 1,
  seed = 2024,
  nsim = 1000,
  test.method = "dunnett",
  approach = "disjoint",
  hazardC = 0.1,
  hazardT = c(0.1, 0.1, 0.1),
  orrC = 0.2,
  orrT = c(0.2, 0.2, 0.2),
  rho = 0.7,
  dropoutC = 0,
  dropoutT = c(0, 0, 0),
  accrual_rate_stage1 = 25,
  accrual_rate_stage2 = 25,
  n1_per_arm = 50,
  n2_per_arm = 150,
  targetEventsIA_all = 240,
  targetEventsFA_all = 320,
  w = NULL,
  bound_z = c(2.44, 2),
  alpha = 0.025,
  update_bound = TRUE,
  nonselected_max_followup = NULL,
  dose_selection_endpoint = "ORR"
)
}
\arguments{
\item{ith}{ith cluster used for parallel computing}

\item{seed}{random seed for reproducibility}

\item{nsim}{number of replicates}

\item{test.method}{method for getting adjusted p-value: \code{"dunnett"} or \code{"simes"}.}

\item{approach}{approach for combining data:
\code{"disjoint"}: disjoint-subjects approach;
\code{"ii"}: independent incremental approach;
\code{"hybrid"}: disjoint-subjects at IA and independent incremental thereafter.}

\item{hazardC}{hazard rate for control arm}

\item{hazardT}{vector of hazard rates for treatment arms}

\item{orrC}{ORR for control arm}

\item{orrT}{vector of ORRs for control arm}

\item{rho}{the correlation between ORR and survival based on Gaussian copula.}

\item{dropoutC}{dropout hazard rate for control.}

\item{dropoutT}{vector of dropout hazard rates for treatment arms.}

\item{accrual_rate_stage1}{number of subjects enrolled per month for stage 1.}

\item{accrual_rate_stage2}{number of subjects enrolled per month for stage 2.}

\item{n1_per_arm}{total number of subjects per arm in stage 1.}

\item{n2_per_arm}{total number of subjects per arm in stage 2.}

\item{targetEventsIA_all}{target number of events at IA for all subjects from both stage 1 and 2.}

\item{targetEventsFA_all}{target number of events at FA for all subjects from both stage 1 and 2.}

\item{w}{weights parameter for stage 1 data: for independent incremental approach, it will be length of 1;
for the disjoint-subjects approach, it will be length of 2 for IA and FA.}

\item{bound_z}{z-scale rejection boundaries and IA and FA.}

\item{alpha}{Type I error, always one-sided.}

\item{update_bound}{whether to re-calculate the FA rejection boundary using the updated
correlation between IA and FA test statistics.}

\item{nonselected_max_followup}{The final data cutoff time for non-selected arms.
The default \code{NULL} means the non-selected arms will not be followed up after dose selection.}

\item{dose_selection_endpoint}{either "ORR" or "Survival"}
}
\value{
It returns a matrix with each row corresponding to the analysis results for each trial,
where the analysis results include: "Selected dose", "IA time", "IA reject", "FA reject", "FA time",
"Study duration", "Total sample size", "observed correlation", and "actual FA boundary".
}
\description{
Get operating characteristics via simulations for the phase 2/3 design
}
\examples{
res <- getOC_ph23(seed = 24232, nsim=10)
apply(res, 2, mean, rm=TRUE)
}
