% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOC_ph23_ii.R
\name{getOC_ph23_ii}
\alias{getOC_ph23_ii}
\title{Get operating characteristics via simulations for the phase 2/3 design}
\usage{
getOC_ph23_ii(
  ith = 1,
  seed = 2024,
  nsim = 1000,
  test.method = "dunnett",
  hazardC = 0.1,
  hazardT = c(0.1, 0.1, 0.1),
  orrC = 0.2,
  orrT = c(0.2, 0.2, 0.2),
  rho = 0.7,
  dropoutC = 0,
  dropoutT = c(0, 0, 0),
  accrual_rate_stage1 = 25,
  accrual_rate_stage2 = 25,
  n1_per_arm = 50,
  n2_per_arm = 150,
  w = NULL,
  targetEventsIA_all = 240,
  targetEventsFA_all = 320,
  bound_z_IA = 2.44,
  bound_z_FA = 2,
  alpha = 0.025,
  update_bound_FA = TRUE,
  nonselected_max_followup = NULL,
  dose_selection_endpoint = "ORR"
)
}
\arguments{
\item{ith}{ith cluster used for parallel computing}

\item{seed}{random seed for reproducibility}

\item{nsim}{number of replicates}

\item{test.method}{method for getting adjusted p-value: \code{"dunnett"} or \code{"simes"}.}

\item{hazardC}{hazard rate for control arm}

\item{hazardT}{vector of hazard rates for treatment arms}

\item{orrC}{ORR for control arm}

\item{orrT}{vector of ORRs for control arm}

\item{rho}{the correlation between ORR and survival based on Gaussian copula.}

\item{dropoutC}{dropout hazard rate for control.}

\item{dropoutT}{vector of dropout hazard rates for treatment arms.}

\item{accrual_rate_stage1}{number of subjects enrolled per month for stage 1.}

\item{accrual_rate_stage2}{number of subjects enrolled per month for stage 2.}

\item{n1_per_arm}{total number of subjects per arm in stage 1.}

\item{n2_per_arm}{total number of subjects per arm in stage 2.}

\item{w}{weight parameter for stage 1 data.}

\item{targetEventsIA_all}{target number of events at IA for all subjects from both stage 1 and 2.}

\item{targetEventsFA_all}{target number of events at FA for all subjects from both stage 1 and 2.}

\item{bound_z_IA}{z-scale rejection boundary at IA.}

\item{bound_z_FA}{z-scale rejection boundary at FA.}

\item{alpha}{Type I error, always one-sided.}

\item{update_bound_FA}{whether to re-calculate the FA rejection boundary using the updated
correlation between IA and FA test statistics.}

\item{nonselected_max_followup}{The final data cutoff time for non-selected arms.
The default \code{NULL} means the non-selected arms will not be followed up after dose selection.}

\item{dose_selection_endpoint}{either "ORR" or "Survival"}
}
\value{
It returns a matrix with each row corresponding to the analysis results for each trial,
where the analysis results include: "Selected dose", "IA time", "IA rejct", "FA reject", "FA time",
"Study duration", "Total sample size", "observed correlation", and "actual FA boundary".
}
\description{
This function uses the independent incremental approach
instead of the disjoint-subject approach used in \link{getOC_ph23}.
}
\examples{
res <- getOC_ph23_ii(seed = 24232, nsim=10)
apply(res, 2, mean, rm=TRUE)
}
