% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getZstat.R
\name{getZstat}
\alias{getZstat}
\title{Get combination test statistic value using disjoint-subjects approach}
\usage{
getZstat(dat, w = NULL, selected, targetEvents, test.method = "dunnett")
}
\arguments{
\item{dat}{A time-to-event dataset returned from \link{sim_ph23}.}

\item{w}{weight parameter for stage 1 data.}

\item{selected}{The selected dose level.}

\item{targetEvents}{The target number of events from both stages at which analysis is to be made.}

\item{test.method}{The method for getting adjusted p-value: \code{"dunnett"} or \code{"simes"}.}
}
\value{
It returns a list including \code{z} the test statistic, \code{obsEvents_stage1} the
observed number of events from stage 1 and \code{cut_time} the data cut time.
}
\description{
Get combination test statistic value using disjoint-subjects approach
}
\examples{
d <- sim_ph23(n1_per_arm = 20, n2_per_arm = 30)
getZstat(d, w=0.8, selected = 2, targetEvents = 25, test.method="dunnett")
}
