% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getZstats.R
\name{getZstats}
\alias{getZstats}
\title{Get test statistics for a given trial.}
\usage{
getZstats(dat, targetEvents.S)
}
\arguments{
\item{dat}{A time-to-event dataset returned from \link{simu_enrich_trial}.}

\item{targetEvents.S}{The target number of events in S; currently only support length of 2, where
futility or population selection will be given at IA without efficacy testing.}
}
\value{
It returns a list of test statistics used for later adjustments.
}
\description{
Get test statistics for a given trial.
}
\examples{
d <- simu_enrich_trial(n = 200, prop_S = 0.5, duration = 10)
getZstats(d, targetEvents.S = c(28, 70))
}
