% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getZtests_AED.R
\name{getZtests_AED}
\alias{getZtests_AED}
\title{Get adaptive enrichment design test statistics for a given trial.}
\usage{
getZtests_AED(
  object,
  alpha = 0.025,
  HR.Sc.threshold = 1,
  HR.S.threshold = 0.8,
  HR.F.threshold = 0.8
)
}
\arguments{
\item{object}{An object return by \link{getZstats}.}

\item{alpha}{Type I error, always one-sided.}

\item{HR.Sc.threshold}{Hazard ratio threshold for futility of Sc.}

\item{HR.S.threshold}{Hazard ratio threshold for S.}

\item{HR.F.threshold}{Hazard ratio threshold for F.}
}
\value{
It returns a list of test statistics used for later adjustments.
}
\description{
Get adaptive enrichment design test statistics for a given trial.
}
\examples{
d <- simu_enrich_trial(n = 200, prop_S = 0.5, duration = 10)
object <- getZstats(d, targetEvents.S = c(28, 70))
getZtests_AED(object)
}
