% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ph23.R
\name{sim_ph23}
\alias{sim_ph23}
\title{Simulate a seamless phase 2/3 trial}
\usage{
sim_ph23(
  hazardC = 0.1,
  hazardT = c(0.09, 0.08, 0.07),
  orrC = 0.2,
  orrT = c(0.25, 0.3, 0.4),
  rho = 0.7,
  accrual_rate_stage1 = 25,
  accrual_rate_stage2 = 25,
  n1_per_arm = 50,
  n2_per_arm = 150,
  dropoutC = 0,
  dropoutT = c(0, 0, 0)
)
}
\arguments{
\item{hazardC}{hazard rate for control arm}

\item{hazardT}{vector of hazard rates for treatment arms}

\item{orrC}{ORR for control arm}

\item{orrT}{vector of ORRs for control arm}

\item{rho}{the correlation between ORR and survival based on Gaussian copula.}

\item{accrual_rate_stage1}{number of subjects enrolled per month for stage 1.}

\item{accrual_rate_stage2}{number of subjects enrolled per month for stage 2.}

\item{n1_per_arm}{total number of subjects per arm in stage 1.}

\item{n2_per_arm}{total number of subjects per arm in stage 2.}

\item{dropoutC}{dropout hazard rate for control.}

\item{dropoutT}{vector of dropout hazard rates for treatment arms.}
}
\value{
A time-to-event dataset consists of
\itemize{
\item \code{stage}: 1=phase II; 2=phase III.
\item \code{trt}: 0=control; 1=dose 1; 2=dose 2; ...
\item \code{enterTime}: subject entry time relative to the study starting date
\item \code{response}: 0=no response; 1=response
\item \code{survTime}: observed event time
\item \code{event}: 0=censored; 1=event
\item \code{calendarTime}: total time on study, i.e. \code{enterTime} + \code{survTime}.
}
}
\description{
This function will simulate a phase 2/3 trial where patients will be randomized to all arms in stage 1
and only the selected arm and control arm will go into stage 2. Since we don't know the arm selection in
advance, we will simulated patients for all arms in stage 2 but with adjusted accrual rate assuming stage 2
only contains the selected arm and the control.
}
\examples{
d <- sim_ph23(n1_per_arm = 20, n2_per_arm = 30)
head(d)
}
