% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simes.R
\name{simes}
\alias{simes}
\title{Adjusted p value for testing H_J using Simes method}
\usage{
simes(p)
}
\arguments{
\item{p}{A vectory of individual raw p values in J.}
}
\value{
Adjusted p value using simes procedure
}
\description{
This functions provides the adjusted p value for testing a familywise hypothesis H_J based on the p values of individual raw p values in J.
}
\examples{
#Example (1):

p = c(0.01, 0.02, 0.03, 0.013)
simes(p)

}
