% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_enrich_trial.R
\name{simu_enrich_trial}
\alias{simu_enrich_trial}
\title{Simulate an oncology trial with sub-population and full population}
\usage{
simu_enrich_trial(
  n = 100,
  prop_S = 0.5,
  ratio = 1,
  duration = 5,
  hazard_S = c(0.1, 0.07),
  hazard_Sc = c(0.11, 0.088),
  dropout_S = c(0, 0),
  dropout_Sc = c(0, 0),
  w = 1
)
}
\arguments{
\item{n}{total number of subjects for full population.}

\item{prop_S}{proportion for sub-population group S.}

\item{ratio}{randomization ratio r:1, where r refers to treatment arm; for equal randomization, r=1.}

\item{duration}{enrollment duration in months.}

\item{hazard_S}{hazard rates (h_control, h_treatment) for S}

\item{hazard_Sc}{hazard rates (h_control, h_treatment) for Sc}

\item{dropout_S}{dropout hazard rates (h_control, h_treatment) for S}

\item{dropout_Sc}{dropout hazard rates (h_control, h_treatment) for Sc}

\item{w}{weight parameter in cumulative enrollment pattern.}
}
\value{
A time-to-event dataset consists of
\itemize{
\item \code{subgroup}: 1=S; 0=Sc.
\item \code{trt}: 1=treatment; 0=control
\item \code{enterTime}: subject entry time relative to the study starting date
\item \code{survTime}: observed event time
\item \code{event}: 0=censored; 1=event
\item \code{calendarTime}: total time on study, i.e. \code{enterTime} + \code{survTime}.
}
}
\description{
This function will simulate a trial where patients will be equally randomized to treatment and control within sub-population S and its complement sub-population Sc.
}
\examples{
d <- simu_enrich_trial(n = 100, prop_S = 0.5, ratio = 1)
head(d)
}
