% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_single_arm.R
\name{simu_single_arm}
\alias{simu_single_arm}
\title{Simulate a single arm survival data with uniform enrollment distribution}
\usage{
simu_single_arm(n = 100, duration = 5, hazard = 0.1, dropout = 0, w = 1)
}
\arguments{
\item{n}{total number of subjects for full population.}

\item{duration}{enrollment duration in months; must be an integer.}

\item{hazard}{hazard rate}

\item{dropout}{dropout hazard rate}

\item{w}{weight parameter in cumulative enrollment pattern.}
}
\value{
A time-to-event dataset consists of
\itemize{
\item \code{enterTime}: subject entry time relative to the study starting date
\item \code{survTime}: observed event time
\item \code{event}: 0=censored; 1=event
\item \code{calendarTime}: total time on study, i.e. \code{enterTime} + \code{survTime}.
}
}
\description{
Simulate a single arm survival data with uniform enrollment distribution
}
\examples{
d <- simu_single_arm(n = 20, duration = 5)
head(d)
}
