% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_by_date.R
\name{cut_by_date}
\alias{cut_by_date}
\title{Cut a dataset for analysis at a specified date}
\usage{
cut_by_date(data, cut_time)
}
\arguments{
\item{data}{A time-to-event dataset returned from \link{sim_ph23}.}

\item{cut_time}{Date relative to start of randomization (\code{calendarTime} from
input dataset) at which dataset is to be cut off for analysis.}
}
\value{
A data frame ready for survival analysis including columns \code{calendarCutoff},
\code{survTimeCut}, \code{eventCut}.
}
\description{
Cut a dataset for analysis at a specified date
}
\examples{
d <- sim_ph23(n1_per_arm = 20, n2_per_arm = 30)
dcut <- cut_by_date(d, 10)
}
