% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_by_event.R
\name{cut_by_event}
\alias{cut_by_event}
\title{Cut a dataset for analysis at a specified event count}
\usage{
cut_by_event(data, targetEvents)
}
\arguments{
\item{data}{A time-to-event dataset returned from \link{simu_enrich_trial}.}

\item{targetEvents}{Event count at which data cutoff is to be made.}
}
\value{
A data frame ready for survival analysis including columns \code{calendarCutoff},
\code{survTimeCut}, \code{eventCut}.
}
\description{
Cut a dataset for analysis at a specified event count
}
\examples{
d <- simu_enrich_trial(n = 100, prop_S = 0.5, ratio = 1)
dcut <- cut_by_event(d, 10)
}
