% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getZstat.R
\name{getZstat}
\alias{getZstat}
\title{Get combination test statistic value}
\usage{
getZstat(dat, w, selected = 1, targetEvents, test.method = "dunnett")
}
\arguments{
\item{dat}{A time-to-event dataset returned from \link{sim_ph23}.}

\item{w}{Weights (w1, w2) used in combination test.}

\item{selected}{The selected dose level.}

\item{targetEvents}{The target number of events from both stages at which analysis is to be made.}

\item{test.method}{The method for getting adjusted p-value: \code{"dunnett"} or \code{"simes"}.}
}
\value{
It returns a list including \code{z} the test statistic, \code{obsEvents_stage1} the
observed number of events from stage 1 and \code{cut_time} the data cut time.
}
\description{
Get combination test statistic value
}
\examples{
d <- sim_ph23(n1_per_arm = 20, n2_per_arm = 30)
getZstat(d, w=c(0.8, sqrt(1-0.8^2)), selected = 2, targetEvents = 25, test.method="dunnett")
}
